Make Clock Time

The function combines a date and a clock time into a POSIXct vector. Clock time can be supplied as decimal hours, for example 13.5, or as a character time, for example "13:30" or "13:30:00".

Usage

MakeClockTime(
  date,
  time,
  tz = "UTC",
  date_formats = c("%m/%d/%y", "%m/%d/%Y", "%Y-%m-%d"),
  invalid = c("NA", "error")
)

Arguments

date

Date vector. Dates can be supplied as Date, POSIXt, or character values.

time

Numeric or character vector. Clock time in decimal hours, "HH:MM", or "HH:MM:SS" format.

tz

Character string. Time zone passed to as.POSIXct().

date_formats

Character vector. Date formats used to parse character dates.

invalid

Character string. If invalid = "NA", invalid clock times are returned as NA. If invalid = "error", invalid clock times throw an error.

Value

Returns a POSIXct vector.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

MakeClockTime(
  date = "2020-01-02",
  time = 13.5,
  tz = "America/New_York"
)
#> [1] "2020-01-02 13:30:00 EST"

MakeClockTime(
  date = "01/02/20",
  time = "13:30",
  tz = "America/New_York"
)
#> [1] "2020-01-02 13:30:00 EST"